faskes-service/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
kafka-python==2.0.2
